fix(metric_alerts): Fix bug where start bucket can be missing on a very new alert.
This fixes a bug where the start bucket can be missing on a new alert. This typically happens on an
alert with a large time window such as an hour. Basically if the start date is after the last bucket
on the graph, it ends up filtered out by the zerofill code.

An example:
We have an alert rule with a time window of 1 hour.
An alert starts at 11:34pm.
We fetch the data, and only fetch up to bucket 11pm since 12am hasn't happened yet.
We have code that fetches the 11:34pm bucket as well, so that is included. However, the end date of
our result ends up being 11pm still, and so when we do zerofill it excludes that bucket.

This fixes the code to include the max of the stat range or the incident start date. It also
modifies the zerofill code to include any leftover buckets. This is necessary because if the start
date bucket is the last bucket, it'll end up filtered off since we round the end date down to
nearest rollup below it.